Write segment closest to target to disk
The worm.h functions will now write the two points on the contour
corresponding the segment that is just in front of the target. This is
to be used for measuring the width of the worm at target.
Prints location of points to .yaml file and to output video.
Note that this had not been debugged yet, only added and compiled.